backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
openai
anthropic
pydantic
sse-starlette
